[refactor] 자바스크립트 모듈 분리
