+++ index.html
@@ -80,6 +80,6 @@
         class="btn unsub-btn hide"
       />
     </div>
-    <script type="module" src="./script.js"></script>
+    <script type="module" src="./app.js"></script>
   </body>
 </html>
