+++ script/grid-view.js
@@ -1,0 +1,55 @@
+
+const gridContainer = document.querySelector(".grid-box");
+const pressCover = document.querySelector(".press-cover");
+const subBtn = document.querySelector(".sub-btn");
+const unsubBtn = document.querySelector(".unsub-btn");
+
+
+let pressIdxArray = Array.from(Array(96).keys()); // create array of consecutive numbers [0...95] - to be used in drawPress()
+let subscribedPress = Array.from(Array(48).keys());  // array of subscribed press IDs
+
+function shuffleArray(arr){
+    for (let i = arr.length - 1; i>0;i--){
+        const j = Math.floor(Math.random()*(i+1));
+        [arr[i], arr[j]] = [arr[j], arr[i]];
+    }
+    return arr;
+}
+
+function checkSubscription(item){
+    if (subscribedPress.includes(parseInt(item.getAttribute("index")))){
+        subBtn.classList.add("hide");
+        unsubBtn.classList.remove("hide");
+    } else {
+        subBtn.classList.remove("hide");
+        unsubBtn.classList.add("hide");
+    }
+}
+function listenPressHover(){
+    const pressItems = document.querySelectorAll(".pressItem");
+    pressItems.forEach((item)=>{
+        item.addEventListener("mouseover",()=>{
+            checkSubscription(item);
+            pressCover.classList.remove("hidden");
+            item.appendChild(pressCover);
+        })
+    })
+    gridContainer.addEventListener("mouseout", () => {
+        pressCover.classList.add("hidden");
+    })
+}
+function drawPress(idx){
+    gridContainer.innerHTML = "";
+    let shuffledArray = shuffleArray(pressIdxArray);
+
+    for (let i=24*idx;i<24*(idx+1);i++){
+        gridContainer.innerHTML += `
+            <li class="pressItem" index=${shuffledArray[i]+1}>
+                <img src="./asset/logo/light/img${shuffledArray[i]+1}.svg" />
+            </li>
+        `
+    }
+    listenPressHover();
+}
+
+export {drawPress}
